Fix Chef DoorLock lock/unlock functions didn't overide weak functions in door-lock-server.cpp (#20538)
* Fix Chef DoorLock lock/unlock functions didn't overide weak functions in door-lock-server.cpp

* Fix restyle issue in src/app/chip_data_model.gni

* Restyled by whitespace

* Restyled by clang-format

* Restyled by gn
