+++ examples/chef/common/stubs.cpp
@@ -4,13 +4,18 @@
 #include <app-common/zap-generated/callback.h>
 #include <app-common/zap-generated/cluster-id.h>
 #include <app-common/zap-generated/command-id.h>
+#include <app/clusters/door-lock-server/door-lock-server.h>
 
-bool emberAfPluginDoorLockOnDoorLockCommand(chip::EndpointId endpointId, chip::Optional<chip::ByteSpan> pinCode)
+bool emberAfPluginDoorLockOnDoorLockCommand(chip::EndpointId endpointId, const chip::Optional<chip::ByteSpan> & pinCode,
+                                            chip::app::Clusters::DoorLock::DlOperationError & err)
 {
+    err = DlOperationError::kUnspecified;
     return true;
 }
 
-bool emberAfPluginDoorLockOnDoorUnlockCommand(chip::EndpointId endpointId, chip::Optional<chip::ByteSpan> pinCode)
+bool emberAfPluginDoorLockOnDoorUnlockCommand(chip::EndpointId endpointId, const chip::Optional<chip::ByteSpan> & pinCode,
+                                              chip::app::Clusters::DoorLock::DlOperationError & err)
 {
+    err = DlOperationError::kUnspecified;
     return true;
 }
