+++ src/app/clusters/door-lock-server/door-lock-server.cpp
@@ -3603,155 +3603,6 @@ void MatterDoorLockPluginServerInitCallback()
 void MatterDoorLockClusterServerAttributeChangedCallback(const app::ConcreteAttributePath & attributePath) {}
 
 // =============================================================================
-// 'Default' callbacks for cluster commands
-// =============================================================================
-
-bool __attribute__((weak))
-emberAfPluginDoorLockOnDoorLockCommand(chip::EndpointId endpointId, const Optional<ByteSpan> & pinCode, DlOperationError & err)
-{
-    err = DlOperationError::kUnspecified;
-    return false;
-}
-
-bool __attribute__((weak))
-emberAfPluginDoorLockOnDoorUnlockCommand(chip::EndpointId endpointId, const Optional<ByteSpan> & pinCode, DlOperationError & err)
-{
-    err = DlOperationError::kUnspecified;
-    return false;
-}
-
-// =============================================================================
-// 'Default' pre-change callbacks for cluster attributes
-// =============================================================================
-
-chip::Protocols::InteractionModel::Status __attribute__((weak))
-emberAfPluginDoorLockOnLanguageChange(chip::EndpointId EndpointId, chip::CharSpan newLanguage)
-{
-    return chip::Protocols::InteractionModel::Status::Success;
-}
-
-chip::Protocols::InteractionModel::Status __attribute__((weak))
-emberAfPluginDoorLockOnAutoRelockTimeChange(chip::EndpointId EndpointId, uint32_t newTime)
-{
-    return chip::Protocols::InteractionModel::Status::Success;
-}
-
-chip::Protocols::InteractionModel::Status __attribute__((weak))
-emberAfPluginDoorLockOnSoundVolumeChange(chip::EndpointId EndpointId, uint8_t newVolume)
-{
-    return chip::Protocols::InteractionModel::Status::Success;
-}
-
-chip::Protocols::InteractionModel::Status __attribute__((weak))
-emberAfPluginDoorLockOnOperatingModeChange(chip::EndpointId EndpointId, uint8_t newMode)
-{
-    return chip::Protocols::InteractionModel::Status::Success;
-}
-
-chip::Protocols::InteractionModel::Status __attribute__((weak))
-emberAfPluginDoorLockOnEnableOneTouchLockingChange(chip::EndpointId EndpointId, bool enable)
-{
-    return chip::Protocols::InteractionModel::Status::Success;
-}
-
-chip::Protocols::InteractionModel::Status __attribute__((weak))
-emberAfPluginDoorLockOnEnablePrivacyModeButtonChange(chip::EndpointId EndpointId, bool enable)
-{
-    return chip::Protocols::InteractionModel::Status::Success;
-}
-
-chip::Protocols::InteractionModel::Status __attribute__((weak))
-emberAfPluginDoorLockOnWrongCodeEntryLimitChange(chip::EndpointId EndpointId, uint8_t newLimit)
-{
-    return chip::Protocols::InteractionModel::Status::Success;
-}
-
-chip::Protocols::InteractionModel::Status __attribute__((weak))
-emberAfPluginDoorLockOnUserCodeTemporaryDisableTimeChange(chip::EndpointId EndpointId, uint8_t newTime)
-{
-    return chip::Protocols::InteractionModel::Status::Success;
-}
-
-chip::Protocols::InteractionModel::Status __attribute__((weak))
-emberAfPluginDoorLockOnUnhandledAttributeChange(chip::EndpointId EndpointId, const chip::app::ConcreteAttributePath & attributePath,
-                                                EmberAfAttributeType attrType, uint16_t attrSize, uint8_t * attrValue)
-{
-    return chip::Protocols::InteractionModel::Status::Success;
-}
-
-// =============================================================================
-// Users and credentials access callbacks
-// =============================================================================
-
-bool __attribute__((weak))
-emberAfPluginDoorLockGetUser(chip::EndpointId endpointId, uint16_t userIndex, EmberAfPluginDoorLockUserInfo & user)
-{
-    return false;
-}
-
-bool __attribute__((weak))
-emberAfPluginDoorLockSetUser(chip::EndpointId endpointId, uint16_t userIndex, chip::FabricIndex creator, chip::FabricIndex modifier,
-                             const chip::CharSpan & userName, uint32_t uniqueId, DlUserStatus userStatus, DlUserType usertype,
-                             DlCredentialRule credentialRule, const DlCredential * credentials, size_t totalCredentials)
-{
-    return false;
-}
-
-bool __attribute__((weak))
-emberAfPluginDoorLockGetCredential(chip::EndpointId endpointId, uint16_t credentialIndex, DlCredentialType credentialType,
-                                   EmberAfPluginDoorLockCredentialInfo & credential)
-{
-    return false;
-}
-
-bool __attribute__((weak))
-emberAfPluginDoorLockSetCredential(chip::EndpointId endpointId, uint16_t credentialIndex, chip::FabricIndex creator,
-                                   chip::FabricIndex modifier, DlCredentialStatus credentialStatus, DlCredentialType credentialType,
-                                   const chip::ByteSpan & credentialData)
-{
-    return false;
-}
-
-DlStatus __attribute__((weak)) emberAfPluginDoorLockGetSchedule(chip::EndpointId endpointId, uint8_t weekdayIndex,
-                                                                uint16_t userIndex, EmberAfPluginDoorLockWeekDaySchedule & schedule)
-{
-    return DlStatus::kFailure;
-}
-
-DlStatus __attribute__((weak)) emberAfPluginDoorLockGetSchedule(chip::EndpointId endpointId, uint8_t yearDayIndex,
-                                                                uint16_t userIndex, EmberAfPluginDoorLockYearDaySchedule & schedule)
-{
-    return DlStatus::kFailure;
-}
-
-DlStatus __attribute__((weak))
-emberAfPluginDoorLockGetSchedule(chip::EndpointId endpointId, uint8_t holidayIndex, EmberAfPluginDoorLockHolidaySchedule & schedule)
-{
-    return DlStatus::kFailure;
-}
-
-DlStatus __attribute__((weak))
-emberAfPluginDoorLockSetSchedule(chip::EndpointId endpointId, uint8_t weekdayIndex, uint16_t userIndex, DlScheduleStatus status,
-                                 DlDaysMaskMap daysMask, uint8_t startHour, uint8_t startMinute, uint8_t endHour, uint8_t endMinute)
-{
-    return DlStatus::kFailure;
-}
-
-DlStatus __attribute__((weak))
-emberAfPluginDoorLockSetSchedule(chip::EndpointId endpointId, uint8_t yearDayIndex, uint16_t userIndex, DlScheduleStatus status,
-                                 uint32_t localStartTime, uint32_t localEndTime)
-{
-    return DlStatus::kFailure;
-}
-
-DlStatus __attribute__((weak))
-emberAfPluginDoorLockSetSchedule(chip::EndpointId endpointId, uint8_t holidayIndex, DlScheduleStatus status,
-                                 uint32_t localStartTime, uint32_t localEndTime, DlOperatingMode operatingMode)
-{
-    return DlStatus::kFailure;
-}
-
-// =============================================================================
 // Timer callbacks
 // =============================================================================
